pyproject.toml
[project]
name = "pdf_lab"
version = "0.1.0"
description = "Gradio app to extract text and images from pdf"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.34.2",
    "llama-cloud-services>=0.6.35",
    "pymupdf>=1.26.1",
    "python-dotenv>=1.1.0",
]
